refactor(test): split executor + service npc_ai tests under cap
Tier 3 of #529. Both files exceeded the 700-line hard cap from
CLAUDE.md and are codecov-ignored, so these are pure mechanical
moves grouped by theme — no test-logic changes.

executor/tests.rs (1428) -> executor/tests/:
  mod.rs (fixtures + re-exports), stats, inventory_counter,
  teleport, mission, npc_state, negative_logging.

service/tests/npc_ai.rs (1377) -> service/tests/npc_ai/:
  mod.rs (fixtures: make_ai_fixture / make_aggression_fixture /
  seed_default_ability / seed_target_with_threat), state_machine,
  aggression, selector, ability_range.

make_aggression_fixture stays pub(super) so the sibling
npc_ai_auto_aggro module keeps resolving it via
super::npc_ai::make_aggression_fixture.

#[test] counts preserved: executor 27 -> 27, npc_ai 28 -> 28.
